[NFC] Move generic contexts and metadata references into separate headers
